+++ navigate/core/node.py
@@ -14,7 +14,6 @@ class Node(TypeCheckMixin):
 
         # internal variables -------------------------------------------------------------------------------------------
         self.allow_dates_in_table = False  # whether added tables can contain dates
-        self.just_copied = False  # whether node was just copied (used in Parser)
         self.command_references = []  # CommandReference queue executed by the Parser
 
     def __repr__(self):